refactor: decouple suppliers directory logic, state, and UI components
