+++ app/suppliers/[id]/page.tsx
@@ -7,7 +7,6 @@ import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/com
 import { Badge } from '@/components/ui/badge'
 import { Button } from '@/components/ui/button'
 import {
-  Building2,
   Mail,
   MapPin,
   Phone,
