+++ app/suppliers/suppliers-directory.tsx
@@ -1,10 +1,9 @@
 'use client'
 
-import { useEffect, useState, useMemo } from 'react'
+import { useEffect, useState } from 'react'
 import Link from 'next/link'
 import { createClient } from '@/lib/supabase/client'
 import { Navigation } from '@/components/navigation'
-import { Badge } from '@/components/ui/badge'
 import { Input } from '@/components/ui/input'
 import { Button } from '@/components/ui/button'
 import {
@@ -16,34 +15,18 @@ import {
 } from '@/components/ui/select'
 import {
   Building2,
-  Mail,
   Search,
   Package,
-  ArrowRight,
   Globe,
-  Briefcase,
   CheckCircle2,
   X,
 } from 'lucide-react'
 import { LATAM_COUNTRIES, SECTORS, getCountryLabel, getSectorLabel } from '@/lib/constants'
 import { PRODUCT_CATEGORIES, getLocalizedCategoryLabel } from '@/lib/categories'
 import { useI18n } from '@/lib/i18n/provider'
-import { SupplierLogo } from '@/components/suppliers/supplier-logo'
-
-type Supplier = {
-  id: string
-  company_name: string
-  bio: string | null
-  address: string | null
-  phone: string | null
-  email: string
-  categories: string[] | null
-  products: string[] | null
-  verified: boolean
-  country: string | null
-  sector: string | null
-  logo_url: string | null
-}
+import type { Supplier } from '@/lib/suppliers/directory-utils'
+import { useSuppliersDirectoryFilters } from '@/hooks/use-suppliers-directory-filters'
+import { SupplierDirectoryCard } from '@/components/suppliers/supplier-directory-card'
 
 const FILTER_CATEGORY_VALUES = ['all', ...PRODUCT_CATEGORIES.map((c) => c.value)] as const
 
@@ -58,10 +41,6 @@ export default function SuppliersPage() {
   const supabase = createClient()
   const [suppliers, setSuppliers] = useState<Supplier[]>([])
   const [isLoading, setIsLoading] = useState(true)
-  const [searchQuery, setSearchQuery] = useState('')
-  const [selectedCategory, setSelectedCategory] = useState('all')
-  const [selectedCountry, setSelectedCountry] = useState('all')
-  const [selectedSector, setSelectedSector] = useState('all')
 
   useEffect(() => {
     const loadSuppliers = async () => {
@@ -137,44 +116,20 @@ export default function SuppliersPage() {
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
 
-  const filteredSuppliers = useMemo(() => {
-    const q = searchQuery.toLowerCase()
-    return suppliers.filter((s) => {
-      if (
-        q &&
-        !s.company_name?.toLowerCase().includes(q) &&
-        !s.bio?.toLowerCase().includes(q) &&
-        !s.products?.some((p) => p.toLowerCase().includes(q))
-      )
-        return false
-      if (selectedCategory !== 'all' && !s.categories?.includes(selectedCategory)) return false
-      if (selectedCountry !== 'all' && s.country !== selectedCountry) return false
-      if (selectedSector !== 'all' && s.sector !== selectedSector) return false
-      return true
-    })
-  }, [suppliers, searchQuery, selectedCategory, selectedCountry, selectedSector])
-
-  const stats = useMemo(
-    () => ({
-      total: suppliers.length,
-      verified: suppliers.filter((s) => s.verified).length,
-      countries: new Set(suppliers.map((s) => s.country).filter(Boolean)).size,
-    }),
-    [suppliers]
-  )
-
-  const hasActiveFilters =
-    selectedCategory !== 'all' ||
-    selectedCountry !== 'all' ||
-    selectedSector !== 'all' ||
-    searchQuery !== ''
-
-  const clearAll = () => {
-    setSearchQuery('')
-    setSelectedCategory('all')
-    setSelectedCountry('all')
-    setSelectedSector('all')
-  }
+  const {
+    searchQuery,
+    setSearchQuery,
+    selectedCategory,
+    setSelectedCategory,
+    selectedCountry,
+    setSelectedCountry,
+    selectedSector,
+    setSelectedSector,
+    filteredSuppliers,
+    stats,
+    hasActiveFilters,
+    clearAll,
+  } = useSuppliersDirectoryFilters(suppliers)
 
   return (
     <div className="flex min-h-screen flex-col">
@@ -386,89 +341,7 @@ export default function SuppliersPage() {
         ) : (
           <div className="grid gap-6 md:grid-cols-2 lg:grid-cols-3">
             {filteredSuppliers.map((supplier) => (
-              <Link
-                key={supplier.id}
-                href={`/suppliers/${supplier.id}`}
-                className="group flex flex-col rounded-2xl border-2 border-border bg-card transition-all hover:border-primary/50 hover:shadow-md focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring"
-              >
-                {/* Card header */}
-                <div className="flex-1 p-5">
-                  <div className="mb-3 flex items-start justify-between gap-2">
-                    <SupplierLogo
-                      logoUrl={supplier.logo_url}
-                      companyName={supplier.company_name}
-                      size="md"
-                      className="group-hover:bg-primary/10"
-                    />
-                    <div className="flex flex-wrap items-center justify-end gap-1.5">
-                      {supplier.verified && (
-                        <span className="inline-flex items-center gap-1 rounded-full bg-success/10 px-2 py-0.5 text-[11px] font-semibold text-success ring-1 ring-success/20">
-                          <CheckCircle2 className="h-3 w-3" aria-hidden />
-                          {t('suppliersPage.verified')}
-                        </span>
-                      )}
-                      {supplier.categories?.[0] && (
-                        <Badge variant="outline" className="text-xs">
-                          {categoryLabel(supplier.categories[0])}
-                        </Badge>
-                      )}
-                    </div>
-                  </div>
-
-                  <h3 className="mb-1 text-base font-bold leading-snug transition-colors group-hover:text-primary">
-                    {supplier.company_name}
-                  </h3>
-
-                  {(supplier.sector || supplier.country) && (
-                    <div className="mb-2 flex flex-wrap items-center gap-x-3 gap-y-0.5 text-xs text-muted-foreground">
-                      {supplier.sector && (
-                        <span className="flex items-center gap-1">
-                          <Briefcase className="h-3 w-3 shrink-0" aria-hidden />
-                          {sectorLabel(supplier.sector)}
-                        </span>
-                      )}
-                      {supplier.country && (
-                        <span className="flex items-center gap-1">
-                          <Globe className="h-3 w-3 shrink-0" aria-hidden />
-                          {countryLabel(supplier.country)}
-                        </span>
-                      )}
-                    </div>
-                  )}
-
-                  {supplier.bio && (
-                    <p className="line-clamp-2 text-sm text-muted-foreground">{supplier.bio}</p>
-                  )}
-                </div>
-
-                {/* Products preview */}
-                {supplier.products && supplier.products.length > 0 && (
-                  <div className="mx-5 mb-4 rounded-xl border border-border bg-muted/30 px-3 py-2">
-                    <p className="text-[10px] font-medium uppercase tracking-wider text-muted-foreground">
-                      {t('suppliersPage.products')}
-                    </p>
-                    <p className="mt-0.5 line-clamp-1 text-sm">
-                      {supplier.products.join(', ')}
-                    </p>
-                  </div>
-                )}
-
-                {/* Email */}
-                {supplier.email && (
-                  <div className="mx-5 mb-4 flex items-center gap-2 text-xs text-muted-foreground">
-                    <Mail className="h-3.5 w-3.5 shrink-0" aria-hidden />
-                    <span className="truncate">{supplier.email}</span>
-                  </div>
-                )}
-
-                {/* CTA */}
-                <div className="p-5 pt-0">
-                  <div className="flex w-full items-center justify-center gap-2 rounded-xl bg-muted/50 py-2.5 text-sm font-semibold transition-colors group-hover:bg-primary/10 group-hover:text-primary">
-                    {t('suppliersPage.viewSupplier')}
-                    <ArrowRight className="h-3.5 w-3.5" aria-hidden />
-                  </div>
-                </div>
-              </Link>
+              <SupplierDirectoryCard key={supplier.id} supplier={supplier} />
             ))}
           </div>
         )}
